Added yaml-controller unit tests.
